+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # HEAD
 
 * Added options for removing or escaping unbalanced/unknown tags
+* Check url for allowed protocolls: http and https for images, http, https, ftp, mailto and xmpp for links
 
 # 0.1.0